+++ server/controllers/podcasts-controller.js
@@ -200,6 +200,12 @@ const PatchPodcasts = async (req, res) => {
  */
 const DeletePodcasts = async (req, res) => {
     const { PodcastId } = req.params;
+    //@ts-ignore
+    const { UserId } = req.user;
+    const Podcast = await ReadOneFromPodcasts(PodcastId);
+    if (Podcast.OrganiserId !== UserId) {
+        return res.status(444).json(AlertBoxObject("Cannot Delete", "You are not the organiser of this podcast"))
+    }
     await Promise.all([
         RemovePodcasts(PodcastId),
         DeleteMembersOfEntity(PodcastId),
@@ -217,8 +223,6 @@ const PodcastInit = (Podcast) => {
             IsAdmin: false,
             CanPostActivity: false,
             CanInviteOthers: false,
-            CanUploadPhoto: false,
-            CanCreateAlbum: false,
             CanUploadVideo: false
         }
     }
